+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to list changes made in each version of the tgw_wordpress cookbook.
 
+# 0.2.0 (12-Dec-2020)
+
+- Configure Nginx to host Wordpress site.
+- Fixed bugs in nginx recipe.
+
 # 0.1.0 (10-Dec-2020)
 
 - Installs base server configuration.